resume, random, poweroffsecs
this is a bunch of my changes, that have kind of gotten piled up....  I'm going to remake them off the newer state of the repo now that the contribution system has changed a little, plus they've made some changes on their end....  keep this all for my own reference while I do so.   also note that I discarded the builds that I produced because I couldn't remember which was for what platform any more, so the firmware files here are out of date.